Refactor homepage settings import paths, grouped app routes
